# docs/swagger.yaml
definitions:
  batch.DeleteRequest:
    properties:
      cards:
        items:
          type: integer
        type: array
      connections:
        items:
          type: integer
        type: array
      id:
        type: integer
      nodes:
        items:
          type: integer
        type: array
      projectId:
        type: integer
    type: object
  batch.DeleteResponse:
    type: object
  batch.SaveRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/graph.DBEventCard'
        type: array
      connections:
        items:
          $ref: '#/definitions/graph.DBConnection'
        type: array
      id:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/graph.DBNode'
        type: array
      projectId:
        type: integer
    type: object
  batch.SaveResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/graph.DBEventCard'
        type: array
      connections:
        items:
          $ref: '#/definitions/graph.DBConnection'
        type: array
      id:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/graph.DBNode'
        type: array
      projectId:
        type: integer
    type: object
  common.IDProject:
    properties:
      id:
        type: integer
      projectId:
        type: integer
    type: object
  endpoint.Container:
    properties:
      map:
        additionalProperties:
          $ref: '#/definitions/endpoint.Endpoint'
        type: object
    type: object
  endpoint.DBEndpoint:
    properties:
      headers:
        type: string
      id:
        type: integer
      module:
        type: string
      projectId:
        type: integer
      query:
        type: string
      uri:
        type: string
    type: object
  endpoint.Endpoint:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      module:
        type: string
      uri:
        type: string
      values:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
  engine.Request:
    properties:
      context:
        additionalProperties: true
        type: object
      event:
        $ref: '#/definitions/graph.DataEvent'
      raw:
        type: object
    type: object
  engine.Response:
    properties:
      error:
        type: string
      noExecutions:
        type: boolean
      response:
        type: object
      scheduled:
        type: boolean
      timeout:
        type: boolean
    type: object
  graph.DBConnection:
    properties:
      graphId:
        type: integer
      id:
        type: integer
      localId:
        description: Placement
        type: integer
      name:
        type: string
      projectId:
        type: integer
      sourceId:
        type: integer
      sourcePort:
        type: string
      targetId:
        type: integer
      targetPort:
        type: string
      ui:
        type: string
    type: object
  graph.DBEventCard:
    properties:
      contextId:
        type: string
      contextType:
        type: string
      graphId:
        type: integer
      httpVote:
        type: integer
      id:
        type: integer
      initiatorId:
        type: string
      initiatorType:
        type: string
      name:
        type: string
      ownerId:
        type: string
      ownerType:
        type: string
      platform:
        type: string
      projectId:
        type: integer
      resourceId:
        type: string
      resourceType:
        type: string
      slidePort:
        type: string
      staticId:
        type: string
      staticType:
        type: string
      targetId:
        description: Placement
        type: integer
      ui:
        type: string
    type: object
  graph.DBGraph:
    properties:
      cards:
        items:
          $ref: '#/definitions/graph.DBEventCard'
        type: array
      connections:
        items:
          $ref: '#/definitions/graph.DBConnection'
        type: array
      counter:
        type: integer
      id:
        type: integer
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/graph.DBNode'
        type: array
      projectId:
        type: integer
      ui:
        type: string
    type: object
  graph.DBNode:
    properties:
      arguments:
        type: string
      function:
        type: string
      graphId:
        type: integer
      id:
        type: integer
      localId:
        description: Placement
        type: integer
      module:
        description: Invocation
        type: string
      name:
        type: string
      projectId:
        type: integer
      ui:
        type: string
    type: object
  graph.DataEvent:
    properties:
      contextId:
        type: string
      contextType:
        type: string
      initiatorId:
        type: string
      initiatorType:
        type: string
      ownerId:
        type: string
      ownerType:
        type: string
      platform:
        type: string
      resourceId:
        type: string
      resourceType:
        type: string
      staticId:
        type: string
      staticType:
        type: string
    type: object
  graph.IDGraph:
    properties:
      graphId:
        type: integer
      id:
        type: integer
      projectId:
        type: integer
    type: object
  search.ActiveGraph:
    properties:
      active:
        items:
          $ref: '#/definitions/graph.DBEventCard'
        type: array
      graph:
        $ref: '#/definitions/graph.DBGraph'
    type: object
  search.FindActiveRequest:
    properties:
      contextId:
        type: string
      contextType:
        type: string
      initiatorId:
        type: string
      initiatorType:
        type: string
      ownerId:
        type: string
      ownerType:
        type: string
      platform:
        type: string
      resourceId:
        type: string
      resourceType:
        type: string
      staticId:
        type: string
      staticType:
        type: string
    type: object
  search.FindActiveResponse:
    properties:
      graphs:
        items:
          $ref: '#/definitions/search.ActiveGraph'
        type: array
    type: object
  storage.ListGraphResponse:
    properties:
      graphs:
        items:
          $ref: '#/definitions/graph.DBGraph'
        type: array
    type: object
info:
  contact: {}
paths:
  /batch:
    delete:
      consumes:
      - application/json
      operationId: BatchDelete
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/batch.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.DeleteResponse'
      summary: deletes all Nodes, Connections and EventCards from the graph
    post:
      consumes:
      - application/json
      operationId: BatchSave
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/batch.SaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/batch.SaveResponse'
      summary: saves new Nodes, Connections and EventCards
  /call:
    post:
      consumes:
      - application/json
      operationId: Call
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/engine.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/engine.Response'
      summary: executes callback
  /connection:
    delete:
      consumes:
      - application/json
      operationId: DeleteConnection
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.IDGraph'
      produces:
      - application/json
      summary: deletes Connection
    post:
      consumes:
      - application/json
      operationId: SaveConnection
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.DBConnection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.DBConnection'
      summary: saves or updates Connection
  /event-card:
    delete:
      consumes:
      - application/json
      operationId: DeleteEventCard
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.IDGraph'
      produces:
      - application/json
      summary: deletes EventCard
    post:
      consumes:
      - application/json
      operationId: SaveEventCard
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.DBEventCard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.DBEventCard'
      summary: saves or updates EventCard
  /find-active:
    post:
      consumes:
      - application/json
      operationId: FindActive
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/search.FindActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/search.FindActiveResponse'
      summary: finds graphs that would activate for event
  /graph:
    delete:
      consumes:
      - application/json
      operationId: DeleteGraph
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.IDGraph'
      produces:
      - application/json
      summary: deletes graph and all related Nodes, Connections and EventCards
    get:
      consumes:
      - application/json
      operationId: GetGraph
      parameters:
      - description: project id
        in: query
        name: projectId
        required: true
        type: integer
      - description: graph id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.DBGraph'
      summary: gets graph with Nodes, Connections and EventCards
    post:
      consumes:
      - application/json
      operationId: SaveGraph
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.DBGraph'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.DBGraph'
      summary: saves or updates graph. Nodes, Connections and EventCards are ignored
  /list-graphs:
    get:
      consumes:
      - application/json
      operationId: ListGraph
      parameters:
      - description: project ids
        in: query
        items:
          type: integer
        name: projectId
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.ListGraphResponse'
      summary: lists graphs
  /list-module:
    get:
      consumes:
      - application/json
      operationId: ListEndpoint
      parameters:
      - description: project ids
        in: query
        items:
          type: integer
        name: projectId
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/endpoint.Container'
      summary: lists endpoints
  /module:
    delete:
      consumes:
      - application/json
      operationId: DeleteEndpoint
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/common.IDProject'
      produces:
      - application/json
      summary: deletes Module
    post:
      consumes:
      - application/json
      operationId: SaveEndpoint
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/endpoint.DBEndpoint'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/endpoint.DBEndpoint'
      summary: saves Module
  /node:
    delete:
      consumes:
      - application/json
      operationId: DeleteNode
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.IDGraph'
      produces:
      - application/json
      summary: deletes Node
    post:
      consumes:
      - application/json
      operationId: SaveNode
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/graph.DBNode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/graph.DBNode'
      summary: saves or updates node
swagger: "2.0"