# pkg/entities/endpoint/orm.go
package endpoint

import (
	"autoflow/pkg/entities/common"
	"gorm.io/datatypes"
	"time"
)

type DBEndpoint struct {
	common.IDProject
	DataEndpoint
}

type DataEndpoint struct {
	Uri     string `json:"uri"`
	Module  string `gorm:"uniq" json:"module"`
	Headers string `json:"headers"`
	Query   string `json:"query"`
}

type DBError struct {
	common.IDProject
	DataError
}

type DataError struct {
	Request    datatypes.JSON `json:"request"`
	Response   datatypes.JSON `json:"response"`
	Error      string
	StatusCode int       `json:"statusCode"`
	CreatedAt  time.Time `json:"createdAt"`
}